Dsp (#1075)
* [INDI common]
Added functions in common routine source.
* Added photon flux calculation
* Added Planck constant and Lumen value #defines
* Updated absolute magnitude and parallax estimations

* [INDIcom] Added relative magnitude calculation

* [INDIcom] Changed log into log10

* [DSP] Added new plugin for Signal processing

This is applicable on BLOBs of INDI::SensorInterface children or
INDI::CCD classes.
Must implement DSP directly into SensorInterface (and CCD if agreed) or
drivers should have this implemented when developers decide?

Features:
- Buffer transformations like DFT or Laplace, Wavelets.
- Convolutions (Lowpass, Highpass filters), downloaded BLOBs from clients
can be used as convolution matrices.
- Primarily written for SensorInterface, can be applied into CCD, for
star or object recognition for example.
- Creates new BLOBs and uploads them to the client separately from the
main BLOB.

TODO:
- Download is possible from KStars (BLOB write-only and FITS/Picture
parsers)
- Cross and auto correlation.

* [DSP] Added BLOB Download in convolution

* [DSP] fixes for spectrograph_simulator.cpp

* [DSP/Sensors] Fixes for Spectrograph

* [Sensors] Fixed StreamManager issue

* [DSP] Added DSP Manager class
Added DSP initial support (not working at the moment) into Sensor
interface and CCD classes

* [DSP] some little corrections

* [DSP] Fixes for streaming into spectrograph simulator

* [DSP][TEST] Fixed DSP for building, needs testing

* [DSP] various fixes

* [DSP] removed unuseful messages

* [DSP] Further fixes for BLOB upload

* [DSP] End of fixes, can merge.

* [DSP] Added Spectrum into Transforms

* [DSP] Fixed runtime issues

* [DSP] added dsp_stats_range_count in libDSP

* [DSP] Added wavelets draft

* [DSP] Changed CCD simulator - added ISNewBLOB forwarding

* [DSP] Errors in circleci

* [DSP] Further circleci errors

* [DSP] Further circleci errors

* [DSP] Updated activation method

* [DSP] updated DFT shift

* [DSP] Removed unused label

+++ examples/tutorial_eight/simple_spectrograph.h
@@ -38,7 +38,7 @@ class SimpleSpectrograph : public INDI::Spectrograph
     bool updateProperties();
 
     // Spectrograph specific functions
-    bool StartIntegration(float duration);
+    bool StartIntegration(double duration);
     bool paramsUpdated(float sr, float freq, float bps, float bw, float gain);
     bool AbortIntegration();
     int SetTemperature(double temperature);
@@ -55,6 +55,6 @@ class SimpleSpectrograph : public INDI::Spectrograph
     // Struct to keep timing
     struct timeval CapStart { 0, 0 };
 
-    float IntegrationRequest { 0 };
-    float TemperatureRequest { 0 };
+    double IntegrationRequest { 0 };
+    double TemperatureRequest { 0 };
 };
